refactor: decompose SettingsModal into tab components (#747)
Extract 6 settings tabs from the 783-line SettingsModal.tsx into focused
components under components/settings/. Each tab component calls the
relevant context hooks directly, keeping SettingsModal as a pure tab
router + layout shell (~150 lines).

New components:
- TypographySettingsTab: font, line height, indent, chars-per-line, auto-save
- VerticalSettingsTab: scroll behavior and sensitivity
- PosHighlightSettingsTab: POS enable toggle, color pickers, live preview
- LintingSettingsTab: thin wrapper around existing LintingSettings component
- SpeechSettingsTab: voice selection, rate, pitch, volume (owns availableVoices state)
- PowerSettingsTab: power save mode and auto-suggest on battery
